Improve bash scripts.
They are even more robust, beautiful and useful.

+++ etc/run_tests.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
 declare -A expected_results
+declare -A errors_by_function
 
 declare comparisons
 declare moves
@@ -12,25 +13,44 @@ declare expected_moves
 function check_order_in()
 {
   local file="$1"
+  local folder="$2"
+  local delete="$3"
+  local errors=0
 
-  for ordered_file in .temp/*.txt; do
+  for ordered_file in "$folder"/*.txt; do
     sort -C "$ordered_file"
 
     if [[ "$?" -ne 0 ]]; then
       printf "$REDCOLOR\n" "Wrong order!"
       printf '%s\n' "While ordering $file, $ordered_file contains wrong results!"
+      errors=$(($errors + 1))
     fi
 
-    rm "$ordered_file"
+    [[ -n $delete ]] && rm "$ordered_file"
   done
+
+  return $errors
 }
 
 function order_tests()
 {
+  local errors=0
+
+  printf "$BLUECOLOR\n" 'Running test [order_tests]'
+  printf "$BLUECOLOR\n" "$SEPARATOR"
+
   for file in order_tests/*.txt; do
+    printf '%s\n' "$file"
+
     ./main.out < "$file"
-    check_order_in $file
+    check_order_in $file ".temp" "DELETE!"
+    errors=$(($errors + $?))
   done
+
+  if [[ $errors -eq 0 ]]; then
+    printf "$GREENCOLOR\n" "Sorting OK"
+  fi
+  printf '\n'
 
   # Removing unused data files
   rm .temp/data/*
@@ -103,7 +123,7 @@ function get_expected_results()
   local func_name="$1"
 
   IFS=','
-  expected=(${expected_results[$func_name]})
+  expected=(${expected_results["$func_name"]})
   expected_comparisons="${expected[0]}"
   expected_moves="${expected[1]}"
   unset IFS
@@ -111,46 +131,64 @@ function get_expected_results()
 
 function data_tests()
 {
+  printf "$BLUECOLOR\n" 'Running test [data_tests]'
+  printf "$BLUECOLOR\n" "$SEPARATOR"
+
   for file in data_tests/*.txt; do
+    printf '%s\n' "$file"
     ./main.out < "$file"
 
     file=$(get_test_file "$file")
     read_expected_for "$file"
-    check_order_in $file
+    check_order_in $file ".temp" "DELETE!"
 
     # data_file is the csv file of each function,
     # meaning that it stores the function name
     for data_file in .temp/data/*.csv; do
       get_results "$data_file"
       func_name=$(get_func_name_from $data_file)
+
+      # Set associative array of errors with the func_name as key and -1 as
+      # value. This is used to assert which functions performed with success
+      errors_by_function["$func_name"]=0
+
       get_expected_results "$func_name"
 
       [[ -z "$expected" ]] && rm $data_file && continue
 
       # Compare these values with the expected values
       if [[ "$comparisons" != "$expected_comparisons" ]]; then
-        printf "$REDCOLOR\n" "Wrong comparisons number!"
+        printf "\n$YELLOWCOLOR\n" "Warning: Wrong comparisons number!"
         printf "$BLUECOLOR" "$func_name"
-        printf '%s' "$ produced wrong number of comparisons in "
+        printf '%s' " produced wrong number of comparisons in "
         printf "$BLUECOLOR\n" "$file"
-        printf '%s\n\n' "Produced $comparisons, when expected $expected_comparisons"
+        printf '%s\n\n' "Produced $comparisons, but expected $expected_comparisons"
+        errors_by_function["$func_name"]=$((${errors_by_function["$func_name"]} + 1))
       fi
 
       if [[ "$moves" != "$expected_moves" ]]; then
-        printf "$REDCOLOR\n" "Wrong movimentations number!"
+        printf "\n$YELLOWCOLOR\n" "Warning: Wrong movimentations number!"
         printf "$BLUECOLOR" "$func_name"
         printf '%s' " produced wrong number of moves in "
         printf "$BLUECOLOR\n" "$file"
-        printf '%s\n\n' "Produced $moves, when expected $expected_moves"
+        printf '%s\n\n' "Produced $moves, but expected $expected_moves"
+        errors_by_function["$func_name"]=$((${errors_by_function["$func_name"]} + 1))
       fi
 
       rm $data_file
     done
   done
+
+  for func_name in "${!errors_by_function[@]}"; do
+    if [[ "${errors_by_function[$func_name]}" -eq 0 ]]; then
+      printf "$func_name $GREENCOLOR\n" "OK"
+    fi
+  done
 }
 
 function perform_tests()
 {
+  cd tests
   mkdir .temp
   mkdir .temp/data
